Разделила код на модули и добавила пустые модули для дальнейшего выполнения ТЗ
