+++ js/comment-length.js
@@ -1,0 +1,4 @@
+const calculateLengthComment = (commentText, maxLengthComment) =>
+  commentText.length <= maxLengthComment;
+
+export {calculateLengthComment};
